Move all cim definitions to custom-surf, make naming of existing cim enums/types consistent
