+++ src/custom-surf/api/index.ts
@@ -1,4 +1,5 @@
 import { BaseApiClient } from "api";
+import { ServiceTicket, ServiceTicketImpactedIMSCircuit, ServiceTicketWithDetails } from "custom/types";
 import { intl } from "locale/i18n";
 import { setFlash } from "utils/Flash";
 import {
@@ -13,12 +14,9 @@ import {
     Organization,
     ServicePortFilterItem,
     ServicePortSubscription,
-    ServiceTicket,
-    ServiceTicketImpactedIMSCircuit,
-    ServiceTicketWithDetails,
 } from "utils/types";
 
-import { CreateTicketPayload } from "../types";
+import { CreateServiceTicketPayload } from "../types";
 
 abstract class CustomApiClientInterface extends BaseApiClient {
     abstract portSubscriptions: (
@@ -55,8 +53,8 @@ abstract class CustomApiClientInterface extends BaseApiClient {
     abstract free_subnets: (subnet: string, netmask: number, prefixlen: number) => Promise<string[]>;
     abstract contacts: (organisationId: string) => Promise<ContactPerson[]>;
     abstract dienstafnameBySubscription: (subscriptionId: string) => Promise<Dienstafname | undefined>;
-    abstract showForms: () => Promise<string[]>;
-    abstract startForm: (formKey: string, userInputs: {}[]) => Promise<any>;
+    abstract cimShowForms: () => Promise<string[]>;
+    abstract cimStartForm: (formKey: string, userInputs: {}[]) => Promise<any>;
     abstract cimTickets: () => Promise<ServiceTicket[]>;
     abstract cimTicketById: (ticket_id: string) => Promise<ServiceTicketWithDetails>;
     abstract cimPatchImpactedObject: (
@@ -208,14 +206,16 @@ export class CustomApiClient extends CustomApiClientInterface {
             Promise.resolve(undefined)
         );
     };
-    showForms = (): Promise<string[]> => {
+
+    cimShowForms = (): Promise<string[]> => {
         return this.fetchJson("surf/cim/forms");
     };
 
-    startForm = (formKey: string, userInputs: {}[]): Promise<{ id: string }> => {
+    cimStartForm = (formKey: string, userInputs: {}[]): Promise<{ id: string }> => {
         return this.postPutJson("surf/cim/forms/" + formKey, userInputs, "post", false, true);
     };
-    createTicket = (ticket: CreateTicketPayload): Promise<{ id: string }> => {
+
+    cimCreateTicket = (ticket: CreateServiceTicketPayload): Promise<{ id: string }> => {
         return this.postPutJson("surf/cim/tickets/", ticket, "post", false, true);
     };
