+++ src/custom-surf/components/cim/ServiceTicketDetailImpactedObjectsStyling.ts
