+++ src/custom-surf/pages/ServiceTickets.tsx
@@ -5,15 +5,16 @@
 
 import { EuiButton, EuiFlexGroup, EuiFlexItem, EuiPage, EuiPanel, EuiSpacer } from "@elastic/eui";
 import ServiceTicketFilter from "custom/components/ServiceTicketFilter";
+import { tableTickets } from "custom/pages/ServiceTicketsStyling";
+import { ServiceTicket, ServiceTicketProcessState } from "custom/types";
 import { intl } from "locale/i18n";
 import debounce from "lodash/debounce";
-import { tableTickets } from "pages/ServiceTicketsStyling";
 import React from "react";
 import { FormattedMessage, WrappedComponentProps, injectIntl } from "react-intl";
 import ScrollUpButton from "react-scroll-up-button";
 import ApplicationContext from "utils/ApplicationContext";
 import { renderStringAsDateTime } from "utils/Lookups";
-import { Filter, ServiceTicket, ServiceTicketProcessState, SortOption } from "utils/types";
+import { Filter, SortOption } from "utils/types";
 import { isEmpty, stop } from "utils/Utils";
 
 type Column = "jira_ticket_id" | "title" | "ticket_state" | "process_state" | "opened_by" | "start_date";
