+++ src/custom-surf/types.ts
@@ -1,19 +1,116 @@
+/** Common */
+
 export interface Filter {
     name: string;
     count: number;
     selected: boolean;
 }
 
-export enum TicketType {
-    PLANNED_WORK = "created",
-    INCIDENT = "running",
+/** CIM */
+
+export interface CreateServiceTicketPayload {
+    end_date?: string;
+    ims_pw_id: string;
+    jira_ticket_id: string;
+    start_date: string;
+    title: string;
+    type: ServiceTicketType;
 }
 
-export interface CreateTicketPayload {
-    start_date: string;
-    end_date?: string;
+export enum ServiceTicketState {
+    INITIAL = "inital",
+    ACTIVE = "active",
+    ABORTED = "aborted",
+    CLOSED = "closed",
+}
+
+// TODO initial is now relating
+export enum ServiceTicketProcessState {
+    // INITIAL = "initial",
+    RELATING = "relating",
+    OPEN = "open",
+    OPEN_RELATED = "open_related",
+    OPEN_ACCEPTED = "open_accepted",
+    UPDATED = "updated",
+    ABORTED = "aborted",
+    CLOSED = "closed",
+}
+
+export interface ServiceTicket {
+    _id: string;
     jira_ticket_id: string;
-    ims_pw_id: string;
-    type: TicketType;
+    opened_by: string;
+    process_state: ServiceTicketProcessState;
+    start_date: string;
+    ticket_state: ServiceTicketState;
     title: string;
+}
+
+export enum ServiceTicketLogType {
+    OPEN = "open",
+    UPDATE = "update",
+    CLOSE = "close",
+}
+
+export interface ServiceTicketLog {
+    entry_time: string;
+    update_nl: string;
+    update_en: string;
+    logtype: ServiceTicketLogType;
+    logged_by: string;
+}
+
+export enum ServiceTicketImpactedObjectImpact {
+    DOWN = "down",
+    NO_IMPACT = "no_impact",
+    REDUCED_REDUNDANCY = "reduced_redundancy",
+}
+
+export interface ServiceTicketImpactedIMSCircuit {
+    ims_circuit_id: number;
+    ims_circuit_name: string;
+    impact: ServiceTicketImpactedObjectImpact;
+    impact_override?: ServiceTicketImpactedObjectImpact;
+    extra_information?: string;
+}
+
+export interface ServiceTicketCustomer {
+    customer_id: string;
+    customer_name: string;
+    customer_abbrev: string;
+}
+
+export interface ServiceTicketContact {
+    name: string;
+    email: string;
+}
+
+export interface ServiceTicketRelatedCustomer {
+    customer: ServiceTicketCustomer;
+    customer_subscription_description?: string;
+    contact: ServiceTicketContact[];
+}
+
+export interface ServiceTicketImpactedObject {
+    subscription_id: string;
+    logged_by: string;
+    ims_circuits: ServiceTicketImpactedIMSCircuit[];
+    owner_customer: ServiceTicketCustomer;
+    subscription_description: string;
+    owner_customer_description?: string;
+    owner_customer_contacts: ServiceTicketContact[];
+    related_customers: ServiceTicketRelatedCustomer[];
+}
+
+export enum ServiceTicketType {
+    PLANNED_WORK = "planned work",
+    INCIDENT = "incident",
+}
+
+export interface ServiceTicketWithDetails extends ServiceTicket {
+    end_date: string;
+    last_update_time: string;
+    type: ServiceTicketType;
+    logs: ServiceTicketLog[];
+    impacted_objects: ServiceTicketImpactedObject[];
 }
